refactor: StoreInfoT /entities로 이동
