+++ src/widgets/join/model/usePostStoreInfo.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query'
 import { storeIdAtom } from '@/shared'
 import { useAtom } from 'jotai'
 import { postStoreInfo } from '../api/postStoreInfo'
-import { StoreInfoT } from '../api/store'
+import { StoreInfoT } from '@/entities'
 
 function usePostStoreInfo(name: string) {
   const [, setStoreId] = useAtom(storeIdAtom)
